+++ src/types/TodoFilter.ts
@@ -1,0 +1,1 @@
+export type TodoFilter = 'all' | 'active' | 'completed';
