+++ src/types/TodoStatus.ts
@@ -1,0 +1,7 @@
+export type TodoStatus =
+  | 'Unable to load todos'
+  | 'Title should not be empty'
+  | 'Unable to add a todo'
+  | 'Unable to delete a todo'
+  | 'Unable to update a todo'
+  | null;
